# doc/openapi.yml
openapi: 3.0.3
info:
  title: SZZ Web Server
  version: 0.0.1
tags:
  - name: szz
paths:
  /szz/variants:
    get:
      tags:
        - szz
      operationId: szz_variants
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    variant_name: 
                      type: string
  /szz/find_bug_commits/:
    post:
      tags:
        - szz
      operationId: find_bug_commits
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - repository_url
                - bugfix_commit_hash
              properties:
                szz_variant:
                  type: string
                repository_url:
                  type: string
                bugfix_commit_hash: 
                  type: array
                  items:
                    type: string
        required: true
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
  /szz/bug_commits/{request_id}:
    get:
      tags:
        - szz
      parameters:
        - name: request_id
          in: path
          schema:
            type: string
          required: true
      operationId: check_result
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  szz_variant:
                    type: string
                  status:
                    type: string
                  result:
                    type: array
                    items:
                      type: object
                      properties:
                        bugfix_commit_hash:
                          type: string
                        bug_commit_hash:
                          type: array
                          items:
                            type: string
  /szz/bug_commits/{request_id}/fix_hash/{commit_hash}:
    get:
      tags:
        - szz
      parameters:
        - name: request_id
          in: path
          schema:
            type: string
          required: true
        - name: commit_hash
          in: path
          schema:
            type: string
          required: true
      operationId: check_result_for_commit
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  szz_variant:
                    type: string
                  status:
                    type: string
                  result:
                    type: array
                    items:
                      type: object
                      properties:
                        bugfix_commit_hash:
                          type: string
                        bug_commit_hash:
                          type: array
                          items:
                            type: string